Spotify

+++ app/src/main/java/com/example/movieplace/ui/movie/MyMoviesRecyclesViewAdapter.kt
@@ -56,7 +56,7 @@ class MyMoviesRecyclesViewAdapter(private val context: Context?) : RecyclerView.
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = movies[position]
         val num = item.vote_average
-        val star = num/2
+        val star = num.toFloat()/2
         holder.rating.rating = star.toFloat()
         if (context != null) {
             Glide.with(context).load(item.img).centerCrop().into(holder.movieImage)
